Basic Enemy AI
*Implemented basic enemy pathing based on set and additive/subractive
rotational changes, timed with a basic variable keeping track of update
cycles passed.
*2 enemies now start from the upper corners and start on a slant towards
the center of the screen, turning away just before meeting. They then
turn so that they are vertical at the edges of the screen and fire 3
shots consecutively in a fan from top to bottom, targeting the player
before moving towards the top of the screen while firing behind them.
*Changed base rotational calculations on bullets and enemies so that
their directional angle is the same as their rotational one.
*Added a multitude of variables and arrays, including 10 in the Sprite
class.
*Added a Calibri font to the Content folder Fonts and used it to display
the elapsed time on the upper left corner of the screen.