Bullet Upgrade Stages
Created 4 basic upgrade stages which modify the bullets the player
shoots

Upgrade stages are as follows:
Stage 1 is a single bullet firing upwards, centred on the middle of
player
Stage 2 is two bullets firing side by side, adjacent and centred on the
middle of player
Stage 3 is stage 2 followed by a single bullet centred on the middle of
player 2 frames after the first wave
Stage 4 is two waves of stage 2 bullets followed by a wave of stage 3
bullets, all 2 frames apart

Player firing is now all calculated in the PlayerFire Method, which is
called once each update.

Ways to change the upgrade stage are not yet implemented, the only way
to modify this value is via the upgradeStage variable in the Player
Class.